ECV-591 - Separate result/billing builders
