+++ src/EPR.Calculator.Service.Function.UnitTests/Builder/ResultBuilderTests.cs
@@ -5,7 +5,6 @@ using EPR.Calculator.Service.Function.Builder.Detail;
 using EPR.Calculator.Service.Function.Builder.LaDisposalCost;
 using EPR.Calculator.Service.Function.Builder.Lapcap;
 using EPR.Calculator.Service.Function.Builder.LateReportingTonnages;
-using EPR.Calculator.Service.Function.Builder.Modulation;
 using EPR.Calculator.Service.Function.Builder.OnePlusFourApportionment;
 using EPR.Calculator.Service.Function.Builder.ParametersOther;
 using EPR.Calculator.Service.Function.Builder.PartialObligations;
@@ -22,7 +21,7 @@ namespace EPR.Calculator.Service.Function.UnitTests.Builder;
 
 [TestCategory(TestCategories.ResultBuilder)]
 [TestClass]
-public class CalcResultBuilderTests : TestsFor<CalcResultBuilder>
+public class ResultBuilderTests : TestsFor<ResultBuilder>
 {
     private Mock<ICalcResultDetailBuilder> mockCalcResultDetailBuilder = null!;
     private Mock<ICalcResultParameterOtherCostBuilder> mockCalcResultParameterOtherCostBuilder = null!;
@@ -37,9 +36,7 @@ public class CalcResultBuilderTests : TestsFor<CalcResultBuilder>
     private Mock<IReportedProducerService> mockReportedProducerService = null!;
     private Mock<ISelfManagedConsumerWasteService> mockSelfManagedConsumerWasteService = null!;
     private Mock<IProducerFeesBuilder> mockSummaryBuilder = null!;
-    private Mock<ICalcResultReader> mockCalcResultReader = null!;
     private Mock<ICalcResultWriter> mockCalcResultWriter = null!;
-    private Mock<ICalcResultModulationBuilder> mockModulationBuilder = null!;
 
     protected override void TestInitialize()
     {
@@ -56,9 +53,7 @@ public class CalcResultBuilderTests : TestsFor<CalcResultBuilder>
         mockCalcResultProjectedProducersBuilder = fixture.Freeze<Mock<ICalcResultProjectedProducersBuilder>>();
         mockSelfManagedConsumerWasteService = fixture.Freeze<Mock<ISelfManagedConsumerWasteService>>();
         mockReportedProducerService = fixture.Freeze<Mock<IReportedProducerService>>();
-        mockCalcResultReader = fixture.Freeze<Mock<ICalcResultReader>>();
         mockCalcResultWriter = fixture.Freeze<Mock<ICalcResultWriter>>();
-        mockModulationBuilder = fixture.Freeze<Mock<ICalcResultModulationBuilder>>();
     }
 
     [TestMethod]
@@ -180,44 +175,6 @@ public class CalcResultBuilderTests : TestsFor<CalcResultBuilder>
     }
 
     [TestMethod]
-    public async Task Build_ShouldReturnCalcResult_WithProjectedProducers_Billing()
-    {
-        var runContext = TestDataHelper.BillingRun2026;
-        var mockProducers1 = new List<L1Producer>
-        {
-            new(1, [new ProducerDetail { ProducerId = 1, SubsidiaryId = null }])
-        };
-        var mockProducers2 = new List<L1Producer>
-        {
-            new(2, [new ProducerDetail { ProducerId = 2, SubsidiaryId = null }])
-        };
-
-        var mockCalcResultProjectedProducersH1Data = new Mock<List<CalcResultH1ProjectedProducer>>();
-        var mockCalcResultProjectedProducersH2Data = new Mock<List<CalcResultH2ProjectedProducer>>();
-        var mockCalcResultPartialData = new Mock<List<CalcResultPartialObligation>>();
-
-        mockCalcResultReader.Setup(m => m.ReadH1ProjectedData(runContext.RunId, CancellationToken.None))
-            .ReturnsAsync(mockCalcResultProjectedProducersH1Data.Object);
-        mockCalcResultReader.Setup(m => m.ReadH2ProjectedData(runContext.RunId, CancellationToken.None))
-            .ReturnsAsync(mockCalcResultProjectedProducersH2Data.Object);
-        mockCalcResultReader.Setup(m => m.ReadPartialData(runContext.RunId, CancellationToken.None))
-            .ReturnsAsync(mockCalcResultPartialData.Object);
-
-        var result = await testSubject.BuildAsync(runContext, CancellationToken.None);
-
-
-        Assert.IsNotNull(result);
-        Assert.AreSame(mockCalcResultProjectedProducersH1Data.Object.ToImmutableList(), result.CalcResultProjectedProducers.H1ProjectedProducers);
-        Assert.AreSame(mockCalcResultProjectedProducersH2Data.Object.ToImmutableList(), result.CalcResultProjectedProducers.H2ProjectedProducers);
-        Assert.AreSame(mockCalcResultPartialData.Object.ToImmutableList(), result.CalcResultPartialObligations.PartialObligations);
-
-        mockCalcResultScaledupProducersBuilder.Verify(m => m.ConstructAsync(runContext, It.IsAny<IImmutableList<MaterialDetail>>(), It.IsAny<List<L1Producer>>()), Times.Never);
-        mockCalcResultProjectedProducersBuilder.Verify(m => m.Construct(runContext, It.IsAny<IImmutableList<MaterialDetail>>(), It.IsAny<List<L1Producer>>()), Times.Never);
-        mockCalcResultPartialObligationBuilder.Verify(m => m.ConstructAsync(runContext,It.IsAny<IImmutableList<MaterialDetail>>(), It.IsAny<List<L1Producer>>()), Times.Never);
-        mockCalcResultWriter.Verify(m => m.StoreProducerMaterialPackaging(It.IsAny<List<L1Producer>>(), CancellationToken.None), Times.Never);
-    }
-
-    [TestMethod]
     public async Task Build_ShouldReturnCalcResult_WithScaledUpProducers()
     {
         var runContext = TestDataHelper.CalculatorRun2025;
@@ -256,30 +213,6 @@ public class CalcResultBuilderTests : TestsFor<CalcResultBuilder>
     }
 
     [TestMethod]
-    public async Task Build_ShouldReturnCalcResult_WithScaledUpProducers_Billing()
-    {
-        var runContext = TestDataHelper.BillingRun2025;
-        var mockCalcResultScaledUpData = new Mock<List<CalcResultScaledupProducer>>();
-        var mockCalcResultPartialData = new Mock<List<CalcResultPartialObligation>>();
-
-        mockCalcResultReader.Setup(m => m.ReadScaledData(runContext.RunId, CancellationToken.None))
-            .ReturnsAsync(mockCalcResultScaledUpData.Object);
-        mockCalcResultReader.Setup(m => m.ReadPartialData(runContext.RunId, CancellationToken.None))
-            .ReturnsAsync(mockCalcResultPartialData.Object);
-
-        var result = await testSubject.BuildAsync(runContext, CancellationToken.None);
-
-        Assert.IsNotNull(result);
-        Assert.AreSame(mockCalcResultScaledUpData.Object.ToImmutableList(), result.CalcResultScaledupProducers.ScaledupProducers);
-        Assert.AreSame(mockCalcResultPartialData.Object.ToImmutableList(), result.CalcResultPartialObligations.PartialObligations);
-
-        mockCalcResultScaledupProducersBuilder.Verify(m => m.ConstructAsync(runContext,It.IsAny<IImmutableList<MaterialDetail>>(), It.IsAny<List<L1Producer>>()), Times.Never);
-        mockCalcResultPartialObligationBuilder.Verify(m => m.ConstructAsync(runContext,It.IsAny<IImmutableList<MaterialDetail>>(), It.IsAny<List<L1Producer>>()), Times.Never);
-        mockCalcResultProjectedProducersBuilder.Verify(m => m.Construct(runContext,It.IsAny<IImmutableList<MaterialDetail>>(), It.IsAny<List<L1Producer>>()), Times.Never);
-        mockCalcResultWriter.Verify(m => m.StoreProducerMaterialPackaging(It.IsAny<List<L1Producer>>(), CancellationToken.None), Times.Never);
-    }
-
-    [TestMethod]
     public async Task Build_ShouldReturnCalcResult_Write_ResultFile()
     {
         var runContext = TestDataHelper.CalculatorRun2026;
@@ -290,41 +223,5 @@ public class CalcResultBuilderTests : TestsFor<CalcResultBuilder>
         mockCalcResultWriter.Verify(m => m.StoreProducerFees(runContext.RunId, result.ProducerFees, CancellationToken.None), Times.Once);
         mockCalcResultWriter.Verify(m => m.StoreSmcw(runContext.RunId, result.Smcw!, CancellationToken.None), Times.Once);
         mockCalcResultWriter.Verify(m => m.StoreModulationResult(runContext.RunId, result.CalcResultModulation!, CancellationToken.None), Times.Once);
-    }
-
-    [TestMethod]
-    public async Task Build_ShouldReturnCalcResult_Read_Billing()
-    {
-        var runContext = TestDataHelper.BillingRun2026;
-        var mockCalcResultProjectedProducersH1Data = new Mock<List<CalcResultH1ProjectedProducer>>();
-        var mockCalcResultProjectedProducersH2Data = new Mock<List<CalcResultH2ProjectedProducer>>();
-        var mockCalcResultPartialData = new Mock<List<CalcResultPartialObligation>>();
-        var mockSummary = new Mock<ProducerFees>();
-        var mockSmcw = new Mock<SelfManagedConsumerWaste>();
-        var mockMod = new Mock<ModulationResult>();
-
-        mockCalcResultReader.Setup(m => m.ReadH1ProjectedData(runContext.RunId, CancellationToken.None))
-            .ReturnsAsync(mockCalcResultProjectedProducersH1Data.Object);
-        mockCalcResultReader.Setup(m => m.ReadH2ProjectedData(runContext.RunId, CancellationToken.None))
-            .ReturnsAsync(mockCalcResultProjectedProducersH2Data.Object);
-        mockCalcResultReader.Setup(m => m.ReadPartialData(runContext.RunId, CancellationToken.None))
-            .ReturnsAsync(mockCalcResultPartialData.Object);
-        mockCalcResultReader.Setup(m => m.ReadProducerFees(runContext.RunId, CancellationToken.None))
-            .ReturnsAsync(mockSummary.Object);
-        mockCalcResultReader.Setup(m => m.ReadSmcw(runContext.RunId, CancellationToken.None))
-            .ReturnsAsync(mockSmcw.Object);
-        mockCalcResultReader.Setup(m => m.ReadModulationResult(runContext.RunId, CancellationToken.None))
-            .ReturnsAsync(mockMod.Object);
-
-        var result = await testSubject.BuildAsync(runContext, CancellationToken.None);
-
-        Assert.IsNotNull(result);
-        Assert.AreSame(mockSmcw.Object, result.Smcw);
-        Assert.AreSame(mockMod.Object, result.CalcResultModulation);
-        Assert.AreSame(mockSummary.Object, result.ProducerFees);
-
-        mockSelfManagedConsumerWasteService.Verify(m => m.Calculate(runContext, It.IsAny<IImmutableList<MaterialDetail>>()), Times.Never);
-        mockModulationBuilder.Verify(m => m.ConstructAsync(It.IsAny<RunContext>(), It.IsAny<IImmutableList<MaterialDetail>>(), It.IsAny<CalcResultLaDisposalCostData>(), It.IsAny<SelfManagedConsumerWaste>()), Times.Never);
-        mockSummaryBuilder.Verify(m => m.ConstructAsync(runContext, It.IsAny<IImmutableList<MaterialDetail>>(), It.IsAny<CalcResult>(), It.IsAny<SelfManagedConsumerWaste>()), Times.Never);
     }
 }
